logs, minimums, target time

+++ k8s-nodeflow.sh
@@ -7,7 +7,7 @@ function _log() {
   1>&2 echo "$(date) $*"
 }
 
-drain_every=${1:-600}
+drain_all_in=$1
 drain_nodes_with_this_label=$2
 pods_not_running_by_this_label_prevent_draining=$3
 
@@ -17,21 +17,30 @@ while true; do
   while true; do
     pods_not_running=$(2>/dev/null kubectl get pod --no-headers \
       --all-namespaces \
-      -l "$pods_not_running_by_this_label_prevent_draining" --field-selector=status.phase!=Running)
+      -l "$pods_not_running_by_this_label_prevent_draining" --field-selector=status.phase!=Running) || true
 
     [[ "$pods_not_running" == "" ]] && break
-    _log "draining paused, pods with label '$pods_not_running_by_this_label_prevent_draining' not running:"
-    _log "$pods_not_running"
-    sleep 5
+
+    pods_not_running_count=$(echo "$pods_not_running" | wc -l | xargs) || true
+    _log "draining paused, '$pods_not_running_count' pods with label '$pods_not_running_by_this_label_prevent_draining' not running"
+
+    sleep 10
   done
 
+  node=""
+  nodes_count=-1
+
   while true; do
-    node=$(kubectl get node \
+    nodes=$(kubectl get node \
       --no-headers \
       --sort-by=.metadata.creationTimestamp \
       -l "$drain_nodes_with_this_label" \
-      -o custom-columns=":metadata.name" \
-      | head -n 1)
+      -o custom-columns=":metadata.name") || true
+
+    nodes_count=$(echo "$nodes" | wc -l | xargs) || true
+    node=$(echo "$nodes" | head -n 1) || true
+
+    _log "nodes_count: '$nodes_count' node: '$node'"
 
     [[ "$node" != "" ]] && break
 
@@ -39,24 +48,29 @@ while true; do
     sleep 1
   done
 
+  if [[ "$nodes_count" -lt "2" ]]; then
+    _log "nodes_count too low"
+    sleep 10
+    continue
+  fi
+
   while true; do
+    _log "draining '$node'"
     >/dev/null 2>&1 kubectl drain "$node" --delete-local-data --ignore-daemonsets --force && break
 
     _log "failed to issue drain for '$node'"
     sleep 1
   done
 
-  _log "issued drain for '$node'"
-
   took=$((SECONDS - started_at))
-  remaining=$((drain_every - took))
+  remaining=$((drain_all_in - took))
+  per_iteration_delay=$((remaining / nodes_count))
 
-  _log "took: $took"
-  _log "remaining: $remaining"
+  _log "took: '$took' remaining: '$remaining' per_iteration_delay='$per_iteration_delay'"
 
-  if [[ "$remaining" -lt 1 ]]; then
+  if [[ "$per_iteration_delay" -lt 1 ]]; then
     sleep 1
   else
-    sleep "$remaining"
+    sleep "$per_iteration_delay"
   fi
 done
